tried updating directories, extension broken
