+++ js/pokemonMoveCategories.js
